generalize for non-scalar values
