+++ python_modules/dagster/dagster_tests/general_tests/test_serdes.py
@@ -5,14 +5,13 @@ from enum import Enum
 from typing import AbstractSet, Any, Dict, Mapping, NamedTuple, Optional, Sequence
 
 import pytest
-from dagster._check import CheckError, ParameterCheckError, inst_param, set_param
-from dagster._core.events import AssetKey
+from dagster._check import ParameterCheckError, inst_param, set_param
 from dagster._serdes.errors import DeserializationError, SerdesUsageError, SerializationError
 from dagster._serdes.serdes import (
-    AssetKeyMap,
     EnumSerializer,
     FieldSerializer,
     NamedTupleSerializer,
+    SerializableNonScalarKeyMapping,
     SetToSequenceFieldSerializer,
     UnpackContext,
     WhitelistMap,
@@ -725,38 +724,55 @@ def test_enum_custom_serializer():
     assert deserialized == Foo.RED
 
 
-def test_serialize_mapping_keyed_by_asset_key():
+def test_serialize_non_scalar_key_mapping():
     test_env = WhitelistMap.create()
-    asset_key_map = AssetKeyMap({AssetKey(["a", "a_2"]): 1})
 
-    serialized = serialize_value(asset_key_map, whitelist_map=test_env)
-    assert serialized == '{"__asset_key_map__": {"[\\"a\\", \\"a_2\\"]": 1}}'
-    assert asset_key_map == deserialize_value(serialized, whitelist_map=test_env)
+    @_whitelist_for_serdes(whitelist_map=test_env)
+    class Bar(NamedTuple):
+        color: str
 
+    non_scalar_key_mapping = SerializableNonScalarKeyMapping({Bar("red"): 1})
 
-def test_asset_key_map():
-    asset_key_map = AssetKeyMap({AssetKey(["a", "a_2"]): 1})
-
-    assert len(asset_key_map) == 1
-    assert asset_key_map[AssetKey(["a", "a_2"])] == 1
-    assert list(iter(asset_key_map)) == list(iter([AssetKey(["a", "a_2"])]))
-
-    with pytest.raises(NotImplementedError, match="AssetKeyMap is immutable"):
-        asset_key_map["foo"] = None
-
-    with pytest.raises(CheckError, match="Key in Mapping mismatches type"):
-        AssetKeyMap({"a": 1})
+    serialized = serialize_value(non_scalar_key_mapping, whitelist_map=test_env)
+    assert (
+        serialized
+        == """{"__non_scalar_key_mapping_items__": [[{"__class__": "Bar", "color": "red"}, 1]]}"""
+    )
+    assert non_scalar_key_mapping == deserialize_value(serialized, whitelist_map=test_env)
 
 
-def test_mapping_keyed_by_asset_key_in_named_tuple():
+def test_serializable_non_scalar_key_mapping():
     test_env = WhitelistMap.create()
 
     @_whitelist_for_serdes(test_env)
-    class Foo(NamedTuple("_Foo", [("keyed_by_asset_key", Mapping[AssetKey, int])])):
-        def __new__(cls, keyed_by_asset_key):
-            return super(Foo, cls).__new__(cls, AssetKeyMap(keyed_by_asset_key))
+    class Bar(NamedTuple):
+        color: str
 
-    named_tuple = Foo(keyed_by_asset_key={AssetKey(["a", "a_2"]): 1})
+    non_scalar_key_mapping = SerializableNonScalarKeyMapping({Bar("red"): 1})
+
+    assert len(non_scalar_key_mapping) == 1
+    assert non_scalar_key_mapping[Bar("red")] == 1
+    assert list(iter(non_scalar_key_mapping)) == list(iter([Bar("red")]))
+
+    with pytest.raises(NotImplementedError, match="SerializableNonScalarKeyMapping is immutable"):
+        non_scalar_key_mapping["foo"] = None
+
+
+def test_serializable_non_scalar_key_mapping_in_named_tuple():
+    test_env = WhitelistMap.create()
+
+    @_whitelist_for_serdes(test_env)
+    class Bar(NamedTuple):
+        color: str
+
+    @_whitelist_for_serdes(test_env)
+    class Foo(NamedTuple("_Foo", [("keyed_by_non_scalar", Mapping[Bar, int])])):
+        def __new__(cls, keyed_by_non_scalar):
+            return super(Foo, cls).__new__(
+                cls, SerializableNonScalarKeyMapping(keyed_by_non_scalar)
+            )
+
+    named_tuple = Foo(keyed_by_non_scalar={Bar("red"): 1})
     assert (
         deserialize_value(
             serialize_value(named_tuple, whitelist_map=test_env), whitelist_map=test_env
